Improve bulk-action workflow, add panel to footer
Added panel at the bottom with buttons: select all/none, remove/restore
selected.

This is because on mobile screens, the button at the front is too tiny
to click (it is easier to select). Furthermore, moving the controls
down (including counters) makes it easier to discover how many cookies
one has selected/deleted.