Add work-around for bugzil.la/1318948
The cookies.set/cookies.remove APIs cannot modify private browsing
cookies, so as a work-around we send a HTTP request and use the
webRequest API to modify cookies.

To number of requests is minimized by combining multiple cookies in a
single request where possible. Also, cookies are stripped from these
requests, so the server cannot identify the client, and neither add
new cookies.

So much complexity, just to work around that stupid bug...