Remove cookies by setting expiration date in the past
Use cookies.set with expiration date in the past instead of
cookies.remove, because the latter may affect more than one cookie.

For example, if we have a domain cookie for example.com and a host-only
cookie, then `chrome.cookies.remove` will delete both. There is no way
to state that only one should be deleted.

We fix this by taking the original cookie and only changing the
expiration date (and the value too, just in case).