Move Firefox-specific parts to manifest_firefox.json
Use separate manifest.json, because I want to declare a page_action and
browser_action for Firefox (compat with Firefox for Android), but Chrome
refuses to load the extension if both are present.

No functional change in the generated zip file.

To build for Firefox, the source files are first copied to a separate
directory, and then zipped. The act of copying and zipping are separate,
so that when "web-ext run" is used (which watches files automatically),
that we can quickly update the files without waiting for the zip
operation.

Also added new run-firefox target to Makefile that delegates to web-ext.