Hide immutable "NID" cookie from results
It is not possible to edit the cookie, and trying to support
editing/removal of cookies through manual guidance takes lots of
efforts. Users have shown to not like not being able to remove
cookies, even if told why, so now I am hiding the immutable
"NID" cookie from the SafeBrowsing cookie jar.

This is expected to be a temporary solution, since the patches for
the bug are already ready for review (https://bugzil.la/1381197).