Add cookie "whitelist" feature
TODO:
- Persist whitelist across page loads
- Add documentation to README

This feature is intended to serve as a general replacement for use cases
like #3: With the ability to whitelist cookies and preventing them from
being removed without confirmation, one can use "Select All" and click
on remove. (or filter out the whitelisted cookies so that they cannot be
selected in the first place).