Import form: Disable cancel button when unsupported
Right before an import, there is a very short window (milliseconds)
where the import can be aborted. Test as follows:

    document.getElementById('import-import').click();
    document.getElementById('import-cancel').click();

After the cookie writing request has started, it is no longer possible
to abort the import without ending up in an inconsistent/undefined
state. So disable the cancel button until the import finishes.