Add multi-selection tool for mouse-based workflows
Note: The multi-selection tool is of very limited use on small touch
screen devices, because it is much faster and more intuitive to tap on
individual cookie rows instead of selecting text and then clicking on a
button. Therefore I am only listening to mouse events and not touch
events.