Remove fake-api-snippet.js via Makefile
The fake-api-snippet.js file is not included in the package.
To avoid a file not found error, strip the `<script>` tag from the HTML.
This is currently only done for the Firefox package, if I ever decide to
publish a Chrome extension then the Chrome extension's package script
should be updated too.