Add First-Party domain to UI (#1)
Add an optional First-Party-domain column to the table when first-party
isolation is enabled. If not enabled, the field is still visible in the
edit form.

To save space, the path column has been removed from the table as it
usually contains just "/".

Tested in Firefox 62, and toggled privacy.firstparty.isolate multiple
times at runtime. Also tested in Firefox Nightly.
Also tested as a non-extension with mocked APIs in Chrome, Firefox and
Fennec.