Fix missing months on some days (#29)
Using the current date and changing the month components of the date can
result in incorrect month labels when the current day does not exist in
that month. For example, today is the 30th of March. There is no 30
February, so setting the month to February caused a rollover to the next
month, and resulted in March appearing twice at the expense of February.