Update for Fenix support; drop support for old Firefox
To specify support for Fenix (Firefox for Android), we need to set
gecko_android in browser_specific_settings. But setting this key forces
incompatibility with Firefox 68 and earlier because of a bug that got
fixed in Firefox 69: https://bugzil.la/1542351

So we'll drop code used to support these earlier versions. This enables
us to remove some hacks that existed to support the lowest common
denominator, as well as the permissions needed for it.